// src/types/cloudFormation.d.ts
type DsDynamoDBConfig = {
  type: 'AMAZON_DYNAMODB';
  config: {
    tableName: string | IntrinsictFunction;
    useCallerCredentials?: boolean;
    serviceRoleArn?: string | IntrinsictFunction;
    region: string | IntrinsictFunction;
    iamRoleStatements?: IamStatement[];
    versioned?: boolean;
    deltaSyncConfig?: {
      deltaSyncTableName: string;
      baseTableTTL?: number;
      deltaSyncTableTTL?: number;
    };
  };
};

type DsRelationalDbConfig = {
  type: 'RELATIONAL_DATABASE';
  config: {
    region: string;
    relationalDatabaseSourceType?: 'RDS_HTTP_ENDPOINT';
    serviceRoleArn?: string | IntrinsictFunction;
    dbClusterIdentifier: string | IntrinsictFunction;
    databaseName: string | IntrinsictFunction;
    schema: string;
    awsSecretStoreArn: string | IntrinsictFunction;
    iamRoleStatements?: IamStatement[];
  };
};

type DsElasticSearchConfig = {
  type: 'AMAZON_ELASTICSEARCH';
  config: {
    domain: string;
    region: string | IntrinsictFunction;
    serviceRoleArn?: string | IntrinsictFunction;
    endpoint: string | IntrinsictFunction;
    iamRoleStatements?: IamStatement[];
  };
};

type LambdaConfig =
  | {
      functionName: string;
      functionAlias?: string;
    }
  | {
      lambdaFunctionArn: string | IntrinsictFunction;
    };

type DsLambdaConfig = {
  type: 'AWS_LAMBDA';
  config: {
    serviceRoleArn?: string | IntrinsictFunction;
    iamRoleStatements?: IamStatement[];
  } & LambdaConfig;
};

type DsHttpConfig = {
  type: 'HTTP';
  config: {
    endpoint: string | IntrinsictFunction;
    serviceRoleArn?: string | IntrinsictFunction;
    iamRoleStatements?: IamStatement[];
    authorizationConfig?: {
      authorizationType?: 'AWS_IAM';
      awsIamConfig?: {
        signingRegion: string | IntrinsictFunction;
        signingServiceName?: string | IntrinsictFunction;
      };
    };
  };
};

type DsNone = {
  type: 'NONE';
};

export type DataSource = {
  name: string;
  description?: string;
} & (
  | DsHttpConfig
  | DsDynamoDBConfig
  | DsRelationalDbConfig
  | DsElasticSearchConfig
  | DsLambdaConfig
  | DsNone
);
// ...
export type FnGetAtt = {
  'Fn::GetAtt': string[];
};

export type FnJoin = {
  'Fn::Join': [string, (string | IntrinsictFunction)[]];
};

export type FnRef = {
  Ref: string;
};

export type FnSub = {
  'Fn::Sub': [string, Record<string, string | IntrinsictFunction>];
};

export type IntrinsictFunction = FnGetAtt | FnJoin | FnRef | FnSub;

export type DeltaSyncConfig = {
  BaseTableTTL: number;
  DeltaSyncTableName: string | IntrinsictFunction;
  DeltaSyncTableTTL: number;
};

export type CfnDataSource = {
  Type: 'AWS::AppSync::DataSource';
  Properties: {
    ApiId: string | IntrinsictFunction;
    Name: string | IntrinsictFunction;
    Description?: string;
    Type?:
      | 'AWS_LAMBDA'
      | 'AMAZON_DYNAMODB'
      | 'AMAZON_ELASTICSEARCH'
      | 'AMAZON_OPENSEARCH_SERVICE'
      | 'NONE'
      | 'HTTP'
      | 'RELATIONAL_DATABASE';
    ServiceRoleArn?: string | IntrinsictFunction;
    LambdaConfig?: {
      LambdaFunctionArn: string | IntrinsictFunction;
    };
    DynamoDBConfig?: {
      Versioned: boolean;
      TableName: string | IntrinsictFunction;
      AwsRegion: string | IntrinsictFunction;
      UseCallerCredentials: boolean;
      DeltaSyncConfig?: DeltaSyncConfig;
    };
    ElasticsearchConfig?: {
      AwsRegion: string | IntrinsictFunction;
      Endpoint: string | IntrinsictFunction;
    };
    RelationalDatabaseConfig?: {
      RelationalDatabaseSourceType: 'RDS_HTTP_ENDPOINT' | IntrinsictFunction;
      RdsHttpEndpointConfig?: {
        AwsRegion: string | IntrinsictFunction;
        DbClusterIdentifier: string | IntrinsictFunction;
        DatabaseName: string | IntrinsictFunction;
        Schema: string | IntrinsictFunction;
        AwsSecretStoreArn: string | IntrinsictFunction;
      };
    };
    HttpConfig?: {
      Endpoint: string | IntrinsictFunction;
      AuthorizationConfig?: {
        AuthorizationType?: string;
        AwsIamConfig?: {
          SigningRegion: string | IntrinsictFunction;
          SigningServiceName?: string | IntrinsictFunction;
        };
      };
    };
  };
};